runtime fn selector resolving is working
