+++ packages/fuels-core/src/abi_decoder.rs
@@ -50,8 +50,8 @@ impl ABIDecoder {
             ParamType::B256 => Self::decode_b256(data),
             ParamType::String(length) => Self::decode_string(data, *length),
             ParamType::Array(ref t, length) => Self::decode_array(data, t, *length),
-            ParamType::Struct(props) => Self::decode_struct(data, props),
-            ParamType::Enum(variants) => Self::decode_enum(data, variants),
+            ParamType::Struct(props, _) => Self::decode_struct(data, props),
+            ParamType::Enum(variants, _) => Self::decode_enum(data, variants),
             ParamType::Tuple(types) => Self::decode_tuple(data, types),
             ParamType::Generic(_) => Err(CodecError::InvalidData(
                 "ParamType::Generic is not decodable!".to_string(),
@@ -416,7 +416,7 @@ mod tests {
         let data = [
             0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x1, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x1,
         ];
-        let param_type = ParamType::Struct(vec![ParamType::U8, ParamType::Bool]);
+        let param_type = ParamType::Struct(vec![ParamType::U8, ParamType::Bool], vec![]);
 
         let decoded = ABIDecoder::decode_single(&param_type, &data)?;
 
@@ -435,7 +435,7 @@ mod tests {
         // }
 
         let inner_enum_types = EnumVariants::new(vec![ParamType::U32, ParamType::Bool])?;
-        let types = vec![ParamType::Enum(inner_enum_types.clone())];
+        let types = vec![ParamType::Enum(inner_enum_types.clone(), vec![])];
 
         // "0" discriminant and 42 enum value
         let data = [
@@ -463,10 +463,13 @@ mod tests {
 
         let inner_enum_types = EnumVariants::new(vec![ParamType::B256, ParamType::U32])?;
 
-        let struct_type = ParamType::Struct(vec![
-            ParamType::Enum(inner_enum_types.clone()),
-            ParamType::U32,
-        ]);
+        let struct_type = ParamType::Struct(
+            vec![
+                ParamType::Enum(inner_enum_types.clone(), vec![]),
+                ParamType::U32,
+            ],
+            vec![],
+        );
 
         let enum_discriminant_enc = vec![0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x1];
         let enum_data_enc = vec![0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x30, 0x39];
@@ -506,13 +509,19 @@ mod tests {
         //     b: u8[2],
         // }
 
-        let nested_struct = ParamType::Struct(vec![
-            ParamType::U16,
-            ParamType::Struct(vec![
-                ParamType::Bool,
-                ParamType::Array(Box::new(ParamType::U8), 2),
-            ]),
-        ]);
+        let nested_struct = ParamType::Struct(
+            vec![
+                ParamType::U16,
+                ParamType::Struct(
+                    vec![
+                        ParamType::Bool,
+                        ParamType::Array(Box::new(ParamType::U8), 2),
+                    ],
+                    vec![],
+                ),
+            ],
+            vec![],
+        );
 
         let data = [
             0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0xa, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x1, 0x0,
@@ -549,13 +558,19 @@ mod tests {
         // Sway fn: long_function(Foo,u8[2],b256,str[23])
 
         // Parameters
-        let nested_struct = ParamType::Struct(vec![
-            ParamType::U16,
-            ParamType::Struct(vec![
-                ParamType::Bool,
-                ParamType::Array(Box::new(ParamType::U8), 2),
-            ]),
-        ]);
+        let nested_struct = ParamType::Struct(
+            vec![
+                ParamType::U16,
+                ParamType::Struct(
+                    vec![
+                        ParamType::Bool,
+                        ParamType::Array(Box::new(ParamType::U8), 2),
+                    ],
+                    vec![],
+                ),
+            ],
+            vec![],
+        );
 
         let u8_arr = ParamType::Array(Box::new(ParamType::U8), 2);
         let b256 = ParamType::B256;
@@ -611,7 +626,7 @@ mod tests {
         let data = [
             0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
         ];
-        let struct_type = ParamType::Struct(vec![ParamType::Unit, ParamType::U64]);
+        let struct_type = ParamType::Struct(vec![ParamType::Unit, ParamType::U64], vec![]);
 
         let actual = ABIDecoder::decode_single(&struct_type, &data)?;
 
@@ -624,7 +639,7 @@ mod tests {
     fn enums_with_all_unit_variants_are_decoded_from_one_word() -> Result<(), Error> {
         let data = [0, 0, 0, 0, 0, 0, 0, 1];
         let variants = EnumVariants::new(vec![ParamType::Unit, ParamType::Unit])?;
-        let enum_w_only_units = ParamType::Enum(variants.clone());
+        let enum_w_only_units = ParamType::Enum(variants.clone(), vec![]);
 
         let result = ABIDecoder::decode_single(&enum_w_only_units, &data)?;
 
@@ -637,7 +652,7 @@ mod tests {
     fn out_of_bounds_discriminant_is_detected() -> Result<(), Error> {
         let data = [0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 2];
         let variants = EnumVariants::new(vec![ParamType::U32])?;
-        let enum_type = ParamType::Enum(variants);
+        let enum_type = ParamType::Enum(variants, vec![]);
 
         let result = ABIDecoder::decode_single(&enum_type, &data);
