+++ packages/fuels-core/src/abi_encoder.rs
@@ -655,10 +655,13 @@ mod tests {
         }
          */
 
-        let struct_a_type = ParamType::Struct(vec![
-            ParamType::Enum(deeper_enum_variants.clone()),
-            ParamType::Bool,
-        ]);
+        let struct_a_type = ParamType::Struct(
+            vec![
+                ParamType::Enum(deeper_enum_variants.clone(), vec![]),
+                ParamType::Bool,
+            ],
+            vec![],
+        );
 
         let struct_a_token = Token::Struct(vec![
             Token::Enum(Box::new((1, deeper_enum_token, deeper_enum_variants))),
