+++ packages/fuels-core/src/code_gen/custom_types.rs
@@ -4,7 +4,7 @@ mod utils;
 
 pub use enum_gen::expand_custom_enum;
 pub use struct_gen::expand_custom_struct;
-pub use utils::extract_custom_type_name_from_abi_property;
+pub use utils::{extract_custom_type_name_from_abi_property, param_type_calls, Component};
 
 // Doing string -> TokenStream -> string isn't pretty but gives us the opportunity to
 // have a better understanding of the generated code so we consider it ok.
