+++ packages/fuels-core/src/code_gen/custom_types/enum_gen.rs
@@ -146,7 +146,7 @@ fn enum_tokenizable_impl(
                     };
 
                     let variants = match Self::param_type() {
-                        ParamType::Enum(variants) => variants,
+                        ParamType::Enum(variants, _) => variants,
                         other => panic!("Calling {}::param_type() must return a ParamType::Enum but instead it returned: {}", #enum_ident_stringified, other)
                     };
 
@@ -170,7 +170,7 @@ fn enum_parameterize_impl(
                 #(param_types.push(#param_type_calls);)*
 
                 let variants = EnumVariants::new(param_types).unwrap_or_else(|_| panic!("{} has no variants which isn't allowed!", #enum_ident_stringified));
-                ParamType::Enum(variants)
+                ParamType::Enum(variants, vec![#(#generics::param_type()),*])
             }
         }
     }
