+++ packages/fuels-core/src/code_gen/custom_types/struct_gen.rs
@@ -129,7 +129,7 @@ fn struct_parameterized_impl(
             fn param_type() -> ParamType {
                 let mut types = Vec::new();
                 #( types.push(#param_type_calls); )*
-                ParamType::Struct(types)
+                ParamType::Struct(types, vec![#(#generic_parameters::param_type()),*])
             }
         }
     }
