+++ packages/fuels-core/src/tokenizer.rs
@@ -52,10 +52,10 @@ impl Tokenizer {
                 trimmed_value.into(),
                 *length,
             ))),
-            ParamType::Struct(struct_params) => {
+            ParamType::Struct(struct_params, _) => {
                 Ok(Self::tokenize_struct(trimmed_value, struct_params)?)
             }
-            ParamType::Enum(variants) => {
+            ParamType::Enum(variants, _) => {
                 let discriminant = get_enum_discriminant_from_string(trimmed_value);
                 let value = get_enum_value_from_string(trimmed_value);
