+++ packages/fuels-core/src/types.rs
@@ -1,7 +1,9 @@
-mod array;
 mod bits_256;
+mod byte;
+mod native;
 mod sized_ascii_string;
 
-pub use crate::types::array::*;
 pub use crate::types::bits_256::*;
+pub use crate::types::byte::*;
+pub use crate::types::native::*;
 pub use crate::types::sized_ascii_string::*;
