--- packages/fuels-core/src/types/array.rs
@@ -1,50 +1,0 @@
-use crate::{Parameterize, Token, Tokenizable};
-use fuels_types::errors::Error;
-use fuels_types::param_types::ParamType;
-
-impl<const SIZE: usize, T: Parameterize> Parameterize for [T; SIZE] {
-    fn param_type() -> ParamType {
-        ParamType::Array(Box::new(T::param_type()), SIZE)
-    }
-}
-
-impl<const SIZE: usize, T: Tokenizable> Tokenizable for [T; SIZE] {
-    fn from_token(token: Token) -> Result<Self, Error>
-    where
-        Self: Sized,
-    {
-        let gen_error = |reason| {
-            Error::InvalidData(format!(
-                "While constructing an array of size {SIZE}: {reason}"
-            ))
-        };
-
-        match token {
-            Token::Array(elements) => {
-                let len = elements.len();
-                if len != SIZE {
-                    return Err(gen_error(format!(
-                        "Was given a Token::Array with wrong number of elements: {len}"
-                    )));
-                }
-
-                let detokenized = elements
-                    .into_iter()
-                    .map(Tokenizable::from_token)
-                    .collect::<Result<Vec<T>, _>>()
-                    .map_err(|err| {
-                        gen_error(format!(", not all elements could be detokenized: {err}"))
-                    })?;
-
-                Ok(detokenized.try_into().unwrap_or_else(|_| {
-                    panic!("This should never fail since we're checking the length beforehand.")
-                }))
-            }
-            _ => Err(gen_error(format!("Expected a Token::Array, got {token}"))),
-        }
-    }
-
-    fn into_token(self) -> Token {
-        Token::Array(self.map(Tokenizable::into_token).to_vec())
-    }
-}
