+++ packages/fuels/tests/harness.rs
@@ -12,7 +12,7 @@ use fuels::prelude::{
 use fuels_core::abi_encoder::ABIEncoder;
 use fuels_core::parameters::StorageConfiguration;
 use fuels_core::tx::{Address, Bytes32, StorageSlot};
-use fuels_core::types::Bits256;
+use fuels_core::types::{Bits256, Byte};
 use fuels_core::Tokenizable;
 use fuels_core::{constants::BASE_ASSET_ID, Token};
 use fuels_signers::fuel_crypto::SecretKey;
@@ -24,6 +24,7 @@ use std::str::FromStr;
 
 use fuel_core_interfaces::model::Message;
 use fuel_gql_client::client::schema::message::Message as OtherMessage;
+use fuels_core::code_gen::abigen::Abigen;
 
 /// Note: all the tests and examples below require pre-compiled Sway projects.
 /// To compile these projects, run `cargo run --bin build-test-projects`.
@@ -327,7 +328,7 @@ async fn compile_bindings_byte_input() {
     // `SimpleContract` is the name of the contract
     let contract_instance = SimpleContractBuilder::new(null_contract_id(), wallet).build();
 
-    let call_handler = contract_instance.takes_byte(10u8);
+    let call_handler = contract_instance.takes_byte(Byte(10u8));
 
     let encoded = format!(
         "{}{}",
@@ -4050,4 +4051,16 @@ async fn test_rust_result_can_be_encoded() -> Result<(), Box<dyn std::error::Err
     assert!(response.value);
 
     Ok(())
+}
+
+#[test]
+fn something() {
+    let result = Abigen::new(
+        "MyContract",
+        "tests/test_projects/results/out/debug/results-abi.json",
+    )
+    .unwrap()
+    .expand()
+    .unwrap();
+    std::fs::write("/tmp/some_code.rs", result.to_string()).unwrap();
 }
