+++ packages/fuels/tests/test_projects/revert_transaction_error/src/main.sw
@@ -1,5 +1,7 @@
 contract;
 
+use std::assert::assert;
+
 abi MyContract {
     fn make_transaction_fail(input: u64) -> u64;
 }
@@ -8,9 +10,7 @@ const COUNTER_KEY = 0x0000000000000000000000000000000000000000000000000000000000
 
 impl MyContract for Contract {
     fn make_transaction_fail(input: u64) -> u64{
-        asm(r1: input) {
-            rvrt r1;
-        };
+        assert(false);
         42
     }
 }
